requirements.txt
click
jinja2
mergedeep
pyaml
sh
